fix(alerts): auto channels selection
